setup zoid modal
