+++ __sdk__.js
@@ -2,6 +2,11 @@
 /* eslint import/no-commonjs: 0 */
 
 module.exports = {
+    'messaging': {
+        entry:           './src/messaging',
+        staticNamespace: '__messaging__',
+        automatic:        false
+    },
     'muse': {
         entry:           './src/index',
         staticNamespace: '__muse__',
