+++ package.json
@@ -37,6 +37,7 @@
     "mocha": "^4"
   },
   "dependencies": {
-    "@paypal/sdk-client": "^4.0.1"
+    "@paypal/sdk-client": "^4.0.1",
+    "zoid": "^9.0.21"
   }
 }
