+++ src/messaging.js
@@ -1,0 +1,1 @@
+export { Messaging } from './messaging-component'
